# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ATM Back Office API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /atms:
    get:
      operationId: listAtms
      tags: [atms]
      summary: List all ATMs with current status
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [online, offline, error, low_cash, all]
          required: false
        - in: query
          name: portal
          schema:
            type: string
          required: false
      responses:
        "200":
          description: List of ATMs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Atm"
    post:
      operationId: createAtm
      tags: [atms]
      summary: Create a new ATM record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAtmBody"
      responses:
        "201":
          description: Created ATM
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Atm"
  /atms/{id}:
    get:
      operationId: getAtm
      tags: [atms]
      summary: Get ATM detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: ATM detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtmDetail"
    put:
      operationId: updateAtm
      tags: [atms]
      summary: Update ATM record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAtmBody"
      responses:
        "200":
          description: Updated ATM
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Atm"
    delete:
      operationId: deleteAtm
      tags: [atms]
      summary: Delete ATM record
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /atms/{id}/transactions:
    get:
      operationId: getAtmTransactions
      tags: [atms]
      summary: Get ATM transaction history
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: query
          name: days
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Transaction history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AtmTransaction"

  # ─── Dashboard ──────────────────────────────────────────────────────────────
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List alerts
      parameters:
        - in: query
          name: severity
          schema:
            type: string
            enum: [critical, warning, info, all]
          required: false
        - in: query
          name: resolved
          schema:
            type: boolean
          required: false
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
  /alerts/{id}/resolve:
    put:
      operationId: resolveAlert
      tags: [alerts]
      summary: Resolve an alert
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resolved alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alert-rules:
    get:
      operationId: listAlertRules
      tags: [alerts]
      summary: List alert rules
      responses:
        "200":
          description: Alert rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertRule"
    post:
      operationId: createAlertRule
      tags: [alerts]
      summary: Create alert rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertRuleBody"
      responses:
        "201":
          description: Created alert rule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertRule"
  /alert-rules/{id}:
    delete:
      operationId: deleteAlertRule
      tags: [alerts]
      summary: Delete alert rule
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Bookkeeping ────────────────────────────────────────────────────────────
  /accounts:
    get:
      operationId: listAccounts
      tags: [bookkeeping]
      summary: List financial accounts
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
    post:
      operationId: createAccount
      tags: [bookkeeping]
      summary: Add a financial account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountBody"
      responses:
        "201":
          description: Created account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
  /accounts/{id}:
    put:
      operationId: updateAccount
      tags: [bookkeeping]
      summary: Update account
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountBody"
      responses:
        "200":
          description: Updated account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
    delete:
      operationId: deleteAccount
      tags: [bookkeeping]
      summary: Delete account
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /book-transactions:
    get:
      operationId: listBookTransactions
      tags: [bookkeeping]
      summary: List financial transactions
      parameters:
        - in: query
          name: accountId
          schema:
            type: integer
          required: false
        - in: query
          name: category
          schema:
            type: string
          required: false
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          required: false
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          required: false
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookTransaction"
    post:
      operationId: createBookTransaction
      tags: [bookkeeping]
      summary: Add a financial transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookTransaction"
  /book-transactions/{id}:
    delete:
      operationId: deleteBookTransaction
      tags: [bookkeeping]
      summary: Delete transaction
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Atm:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        locationName:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        serialNumber:
          type: string
        portalSource:
          type: string
          enum: [columbus_data, switch_commerce, atm_transact, manual]
        portalAtmId:
          type: string
        cashCapacity:
          type: number
        currentBalance:
          type: number
        lowCashThreshold:
          type: number
        status:
          type: string
          enum: [online, offline, error, low_cash, unknown]
        lastSynced:
          type: string
          format: date-time
        avgDailyTransactions:
          type: number
        avgDailyDispensed:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - locationName
        - address
        - city
        - state
        - portalSource
        - cashCapacity
        - currentBalance
        - lowCashThreshold
        - status
        - createdAt
    AtmDetail:
      allOf:
        - $ref: "#/components/schemas/Atm"
        - type: object
          properties:
            recentAlerts:
              type: array
              items:
                $ref: "#/components/schemas/Alert"
            fillHistory:
              type: array
              items:
                $ref: "#/components/schemas/FillOrder"
    CreateAtmBody:
      type: object
      properties:
        name:
          type: string
        locationName:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        serialNumber:
          type: string
        portalSource:
          type: string
          enum: [columbus_data, switch_commerce, atm_transact, manual]
        portalAtmId:
          type: string
        cashCapacity:
          type: number
        lowCashThreshold:
          type: number
      required:
        - name
        - locationName
        - address
        - city
        - state
        - portalSource
        - cashCapacity
        - lowCashThreshold
    UpdateAtmBody:
      type: object
      properties:
        name:
          type: string
        locationName:
          type: string
        address:
          type: string
        city:
          type: string
        state:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        serialNumber:
          type: string
        cashCapacity:
          type: number
        lowCashThreshold:
          type: number
        status:
          type: string
          enum: [online, offline, error, low_cash, unknown]
    AtmTransaction:
      type: object
      properties:
        id:
          type: integer
        atmId:
          type: integer
        date:
          type: string
          format: date
        transactionCount:
          type: integer
        totalDispensed:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - atmId
        - date
        - transactionCount
        - totalDispensed
        - createdAt

    # ── Dashboard ──
...
    FillOrder:
      type: object
      properties:
        id:
          type: integer
        atmId:
          type: integer
        atmName:
          type: string
        locationName:
          type: string
        scheduledDate:
          type: string
          format: date
        cashAmount:
          type: number
        daysToFill:
          type: integer
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - atmId
        - atmName
        - locationName
        - scheduledDate
        - cashAmount
        - daysToFill
        - status
        - createdAt
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        atmId:
          type: integer
        atmName:
          type: string
        type:
          type: string
          enum: [low_cash, machine_error, offline, out_of_cash, sync_failed, custom]
        severity:
          type: string
          enum: [critical, warning, info]
        message:
          type: string
        resolved:
          type: boolean
        createdAt:
          type: string
          format: date-time
        resolvedAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - severity
        - message
        - resolved
        - createdAt
    AlertRule:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [low_cash, machine_error, offline, out_of_cash]
        threshold:
          type: number
        severity:
          type: string
          enum: [critical, warning, info]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - severity
        - isActive
        - createdAt
    CreateAlertRuleBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [low_cash, machine_error, offline, out_of_cash]
        threshold:
          type: number
        severity:
          type: string
          enum: [critical, warning, info]
      required: [name, type, severity]

    # ── Bookkeeping ──
    Account:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [checking, savings, credit_card, cash, other]
        institution:
          type: string
        lastFour:
          type: string
        balance:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - institution
        - balance
        - isActive
        - createdAt
    CreateAccountBody:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [checking, savings, credit_card, cash, other]
        institution:
          type: string
        lastFour:
          type: string
        balance:
          type: number
      required: [name, type, institution, balance]
    UpdateAccountBody:
      type: object
      properties:
        name:
          type: string
        balance:
          type: number
        isActive:
          type: boolean
    BookTransaction:
      type: object
      properties:
        id:
          type: integer
        accountId:
          type: integer
        accountName:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - accountId
        - accountName
        - date
        - description
        - amount
        - type
        - category
        - createdAt
    CreateBookTransactionBody:
      type: object
      properties:
        accountId:
          type: integer
        date:
          type: string
          format: date
        description:
          type: string
        amount:
          type: number
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        notes:
          type: string
      required: [accountId, date, description, amount, type, category]
...